Code clean-up: Move debug stuff to debug.c
